Fix exception when collecting inherited OneToOneFields
Add test that for inherited fields regression

This fixes a regression introduced in 0.4.1 - when inherited fields
are collected an exception is raised. This was introduced by a875cde8.

That commit replaces get_all_related_objects() with a suggested
equivalent using the new meta API for Django 1.8+. However, that
code has a bug.

The Django docs make some mention of filtering results from the
new meta API based on the concrete attribute:

> direct can be found by: not field.auto_created or field.concrete
>
> The auto_created check excludes all "forward" and "reverse"
> relations that are created by Django, but this also includes
> AutoField and OneToOneField on proxy models. We avoid filtering
> out these attributes using the concrete attribute.

https://docs.djangoproject.com/en/1.9/ref/models/meta/

There is also the following ticket that confirms the suggested code
in the Django docs has this problem and suggests filering on concrete
too:

https://code.djangoproject.com/ticket/25461